+++ src/Controls/tests/UITests/Tests/Layout/ScrollView/LayoutUITests.EmptyContent.cs
@@ -5,31 +5,11 @@ using UITest.Core;
 namespace Microsoft.Maui.AppiumTests
 {
 	[Category(TestCategory.Layout)]
-	public class ScrollViewEmptyContentUITests : UITest
+	public class ScrollViewEmptyContentUITests : LayoutUITests
 	{
-		const string LayoutGallery = "Layout Gallery";
-
 		public ScrollViewEmptyContentUITests(TestDevice device)
 			: base(device)
 		{
-		}
-
-		protected override void FixtureSetup()
-		{
-			base.FixtureSetup();
-			App.NavigateToGallery(LayoutGallery);
-		}
-
-		protected override void FixtureTeardown()
-		{
-			base.FixtureTeardown();
-			this.Back();
-		}
-
-		[TearDown]
-		public void LayoutUITestTearDown()
-		{
-			this.Back();
 		}
 
 		[Test]
