+++ src/Controls/tests/UITests/Tests/Layout/ScrollView/LayoutUITests.ScrollBarVisibilityUITests.cs
@@ -5,31 +5,11 @@ using UITest.Core;
 namespace Microsoft.Maui.AppiumTests
 {
 	[Category(TestCategory.Layout)]
-	public class ScrollBarVisibilityUITests : UITest
+	public class ScrollBarVisibilityUITests : LayoutUITests
 	{
-		const string LayoutGallery = "Layout Gallery";
-
 		public ScrollBarVisibilityUITests(TestDevice device)
 			: base(device)
 		{
-		}
-
-		protected override void FixtureSetup()
-		{
-			base.FixtureSetup();
-			App.NavigateToGallery(LayoutGallery);
-		}
-
-		protected override void FixtureTeardown()
-		{
-			base.FixtureTeardown();
-			this.Back();
-		}
-
-		[TearDown]
-		public void LayoutUITestTearDown()
-		{
-			this.Back();
 		}
 
 		[Test]
