+++ src/Controls/tests/UITests/Tests/Layout/StackLayout/LayoutUITests.HorizontalStackLayout.cs
@@ -5,33 +5,13 @@ using UITest.Core;
 namespace Microsoft.Maui.AppiumTests
 {
 	[Category(TestCategory.Layout)]
-	public class HorizontalStackLayoutUITests : UITest
+	public class HorizontalStackLayoutUITests : LayoutUITests
 	{
-		const string LayoutGallery = "Layout Gallery";
-
 		public HorizontalStackLayoutUITests(TestDevice device)
 			: base(device)
 		{
 		}
-
-		protected override void FixtureSetup()
-		{
-			base.FixtureSetup();
-			App.NavigateToGallery(LayoutGallery);
-		}
-
-		protected override void FixtureTeardown()
-		{
-			base.FixtureTeardown();
-			this.Back();
-		}
-
-		[TearDown]
-		public void LayoutUITestTearDown()
-		{
-			this.Back();
-		}
-				
+		
 		[Test]
 		[Description("Organizes child views in a horizontal one-dimensional stack")]
 		public void HorizontalStackLayout()
